WIP : clean up WIP : add first MVP api
