+++ sample/build.gradle
@@ -29,5 +29,6 @@ dependencies {
   compile project(path: ':sdk')
   compile fileTree(dir: 'libs', include: ['*.jar'])
   compile 'com.android.support:multidex:1.0.1'
+  compile 'com.jakewharton:butterknife:5.1.1'
   testCompile 'junit:junit:4.12'
 }
