+++ sdk/build.gradle
@@ -37,5 +37,4 @@ dependencies {
     compile 'com.google.android.gms:play-services-location:8.4.0'
     compile 'com.google.android.gms:play-services-maps:8.4.0'
     compile 'de.hdodenhof:circleimageview:2.0.0'
-    compile 'com.jakewharton:butterknife:5.1.1'
 }
