--- sdk/src/main/assets/crashlytics-build.properties
@@ -1,11 +1,0 @@
-#This file is automatically generated by Crashlytics to uniquely
-#identify individual builds of your Android application.
-#
-#Do NOT modify, delete, or commit to source control!
-#
-#Mon Feb 29 17:10:17 EET 2016
-version_name=0.0
-package_name=com.magnet.magnetchat
-build_id=8efca969-1923-473b-8c20-6df0befa2278
-version_code=0
-app_name=Messenger
