--- sdk/src/main/res/values/com_crashlytics_export_strings.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8" standalone="no"?>
-<resources xmlns:tools="http://schemas.android.com/tools">
-<!--
-  This file is automatically generated by Crashlytics to uniquely 
-  identify individual builds of your Android application.
-
-  Do NOT modify, delete, or commit to source control!
--->
-<string tools:ignore="UnusedResources,TypographyDashes" name="com.crashlytics.android.build_id" translatable="false">8efca969-1923-473b-8c20-6df0befa2278</string>
-</resources>
